+++ ios/Classes/InstabugFlutterPlugin.h
@@ -5,16 +5,6 @@
 - (void)handleMethodCall:(FlutterMethodCall *)call result:(FlutterResult)result;
 
 /**
-  * Enables and disables everything related to receiving replies.
-  * @param isEnabled isEnabled 
-  */
-+ (void)setRepliesEnabled:(NSNumber *)isEnabled;
-
-/**
-  * Manual invocation for replies.
-  */
-+ (void)showReplies;
-/**
   * Tells whether the user has chats already or not.
   */
 + (void)hasChats;
@@ -30,13 +20,6 @@
   * @return number of messages that are unread for this user
   */
 + (void)getUnreadRepliesCount;
-
-/**
-  * Enabled/disable chat notification
-  *
-  * @param isEnabled whether chat notification is reburied or not
-  */
-+ (void)setChatNotificationEnabled:(NSNumber *)isEnabled;
 
 + (void)networkLog:(NSDictionary *)networkData;
