+++ ios/Classes/RepliesApiImpl.h
@@ -1,0 +1,2 @@
+@interface RepliesApiImpl : NSObject <RepliesApi>
+@end
